refactor(missions/fe): patchMissionTemplate throws ApiValidationError
createMissionTemplate already throws ApiValidationError; patchMissionTemplate
was throwing a generic Error with a JSON.stringify'd body. Consumers had to
manually JSON.parse error messages to surface them — easy trap for the next
developer adding inline PATCH editing.

Both api functions now share the same error shape. flattenErrorMessage
extracted from CreateMissionPage to a shared export in api.ts so it isn't
duplicated. StaffActionsCard's formatErrorMessage (which JSON.parsed the
error message string) replaced with the recursive flattener operating
directly on fieldErrors.

Closes #584
